+++ dharma_transcriptions/downloads/Como conciliar o princípio da não ação com a ação no mundo_/subtitles.srt
@@ -1,0 +1,280 @@
+1
+00:00:00,000 --> 00:00:23,000
+ Então, esse princípio da não-ação, como colocado aqui, ele corresponde a essa capacidade lucida.
+
+2
+00:00:23,000 --> 00:00:31,000
+ A capacidade lucida não precisa ser acionada, por isso que isso é chamado de não-ação.
+
+3
+00:00:31,000 --> 00:00:39,000
+ Então, não tem uma sensação de que tem algo que seja ligado e desligado.
+
+4
+00:00:39,000 --> 00:00:41,000
+ Não tem uma sensação de esforço.
+
+5
+00:00:41,000 --> 00:00:48,000
+ Então, a capacidade lucida é natural.
+
+6
+00:00:48,000 --> 00:00:52,000
+ Então, quando nós removemos os obstáculos,
+
+7
+00:00:52,000 --> 00:01:01,000
+ ou seja, nós removamos a ação da mente baseada em referenciais construídos.
+
+8
+00:01:01,000 --> 00:01:11,000
+ A lucidez natural se torna a ação natural. Ela surge naturalmente.
+
+9
+00:01:11,000 --> 00:01:16,000
+ Então, nesse sentido é que há uma não-ação.
+
+10
+00:01:16,000 --> 00:01:22,000
+ Não é necessário qualquer tipo de elaboração ou qualquer tipo de esforço.
+
+11
+00:01:22,000 --> 00:01:30,000
+ Isso é uma forma de dizer que não há iluminação, não há sabedoria.
+
+12
+00:01:30,000 --> 00:01:37,000
+ Quando a gente olha para a juna parâmita, não há resultado.
+
+13
+00:01:37,000 --> 00:01:40,000
+ Não há um caminho para atingir isso.
+
+14
+00:01:42,000 --> 00:01:48,000
+ Então, é uma forma de reconhecer essa condição última,
+
+15
+00:01:48,000 --> 00:01:50,000
+ com uma condição não elaborada.
+
+16
+00:01:50,000 --> 00:01:57,000
+ Se a gente tiver qualquer tipo de esforço para obter alguma coisa,
+
+17
+00:01:57,000 --> 00:02:00,000
+ aquilo não é algo natural, aquilo é elaborado.
+
+18
+00:02:00,000 --> 00:02:02,000
+ Isso é aquilo é elaborado.
+
+19
+00:02:03,000 --> 00:02:08,000
+ Então, mais adiante, vamos nos defrontar com a impermanência inevitável.
+
+20
+00:02:08,000 --> 00:02:13,000
+ Aqui o tem um início, tem uma duração, tem um final.
+
+21
+00:02:13,000 --> 00:02:19,000
+ Mas o aspecto da lucidez pertence a dar uma caia, dar mata, não há.
+
+22
+00:02:19,000 --> 00:02:27,000
+ Nem uma artificialidade, nem o suportanto, não tem nenhum momento em que isso não estivesse presente.
+
+23
+00:02:27,000 --> 00:02:29,000
+ Essa realização.
+
+24
+00:02:29,000 --> 00:02:35,000
+ Essa realização não surge como um conjunto de palavras ou um estados mentais específicos.
+
+25
+00:02:35,000 --> 00:02:45,000
+ Ela surge como a base primordial sem conteúdo, insescentemente presente.
+
+26
+00:02:45,000 --> 00:02:52,000
+ Aqui, nós acessamos isso na forma do prática parâmito como uma confiança.
+
+27
+00:02:52,000 --> 00:02:57,000
+ E dependendo dos ensinamentos, nós podemos ouvir isso também como o fé.
+
+28
+00:02:57,000 --> 00:03:02,000
+ Eu não gosto muito à tradução como o fé, porque parece que é uma ação.
+
+29
+00:03:02,000 --> 00:03:05,000
+ Eu tenho que desenvolver uma fé.
+
+30
+00:03:05,000 --> 00:03:14,000
+ Mas essa fé é mais parecida com a presença, a natural presença,
+
+31
+00:03:14,000 --> 00:03:19,000
+ simplesmente repousar sem esforço na natural presença.
+
+32
+00:03:19,000 --> 00:03:23,000
+ Agora, no cotidiano, o que acontece com o nosso?
+
+33
+00:03:23,000 --> 00:03:28,000
+ Nós estamos sob o domínio de o padana.
+
+34
+00:03:28,000 --> 00:03:31,000
+ Nesse momento, nós estamos inevitabilmente presos ao padana.
+
+35
+00:03:31,000 --> 00:03:37,000
+ O padana significa o que? Nós temos que fazer esforço o tempo todo.
+
+36
+00:03:37,000 --> 00:03:43,000
+ Nós estamos, o padana é o nono dos doze alos.
+
+37
+00:03:43,000 --> 00:03:47,000
+ Nós somos equilibristas. Nós estamos constantemente equilibrando as coisas.
+
+38
+00:03:47,000 --> 00:03:52,000
+ Nós estamos constantemente equilibrando.
+
+39
+00:03:52,000 --> 00:03:56,000
+ Não só uma coisa, mas nós estamos equilibrando muitas diferentes coisas.
+
+40
+00:03:56,000 --> 00:04:03,000
+ Tem uma época que usava esse exemplo dos pratos gerando no círculo.
+
+41
+00:04:03,000 --> 00:04:05,000
+ Tenho esses equilibristas.
+
+42
+00:04:05,000 --> 00:04:09,000
+ O ciclo terminaram. Não sei, não tem mais como botar.
+
+43
+00:04:09,000 --> 00:04:11,000
+ Esse exemplo não é isso.
+
+44
+00:04:11,000 --> 00:04:14,000
+ Eles botam mais um prato, mais um prato, mais um prato.
+
+45
+00:04:14,000 --> 00:04:16,000
+ No último, o primeiro jota o quadro ainda.
+
+46
+00:04:16,000 --> 00:04:18,000
+ Ele vai retornar, vai acelerar.
+
+47
+00:04:18,000 --> 00:04:21,000
+ Essa é a nossa estação.
+
+48
+00:04:21,000 --> 00:04:23,000
+ Vamos girar um prato aqui.
+
+49
+00:04:23,000 --> 00:04:26,000
+ Nós estamos constantemente em ação.
+
+50
+00:04:26,000 --> 00:04:29,000
+ Nós estamos num ambiente de não ação.
+
+51
+00:04:29,000 --> 00:04:32,000
+ O mundo condicionado, ele exige isso.
+
+52
+00:04:32,000 --> 00:04:36,000
+ Isso é dos três animais ao segundo.
+
+53
+00:04:36,000 --> 00:04:41,000
+ É uma incessante demanda, uma busca incessante.
+
+54
+00:04:41,000 --> 00:04:44,000
+ O mundo se escam, os galos se escam.
+
+55
+00:04:44,000 --> 00:04:48,000
+ Os seres buscam sempre alguma coisa incessentemente.
+
+56
+00:04:48,000 --> 00:04:51,000
+ Nós não temos como evitar isso.
+
+57
+00:04:51,000 --> 00:04:57,000
+ Porque esse é o ponto, é a forma de ela que nós trabalhamos em permanência.
+
+58
+00:04:57,000 --> 00:05:00,000
+ Então, nós a gente perguntar mais como é que começou isso?
+
+59
+00:05:00,000 --> 00:05:03,000
+ Porque todo mundo tem que se mover o tempo todo.
+
+60
+00:05:03,000 --> 00:05:06,000
+ É isso, porque nós somos equilibristas.
+
+61
+00:05:06,000 --> 00:05:09,000
+ Nós somos seres desequilibrados.
+
+62
+00:05:09,000 --> 00:05:12,000
+ Então, nós estamos constantemente em rei que librando.
+
+63
+00:05:12,000 --> 00:05:16,000
+ Não há solução.
+
+64
+00:05:16,000 --> 00:05:19,000
+ Então, a não ação é uma descoberta.
+
+65
+00:05:19,000 --> 00:05:23,000
+ Uma descoberta dentro do mundo da ação.
+
+66
+00:05:23,000 --> 00:05:27,000
+ É alguma coisa extraordinária.
+
+67
+00:05:27,000 --> 00:05:32,000
+ Os mestres de diferentes tradições, eles também descobriram isso.
+
+68
+00:05:32,000 --> 00:05:34,000
+ Não são os mestres do dia.
+
+69
+00:05:34,000 --> 00:05:36,000
+ Eles falam da não ação.
+
+70
+00:05:36,000 --> 00:05:40,000
+ Algumas mestres, algumas tradições.
+
