+++ dharma_transcriptions/downloads/Como conciliar o princípio da não ação com a ação no mundo_/transcription.txt
@@ -1,0 +1,1 @@
+ Então, esse princípio da não-ação, como colocado aqui, ele corresponde a essa capacidade lucida. A capacidade lucida não precisa ser acionada, por isso que isso é chamado de não-ação. Então, não tem uma sensação de que tem algo que seja ligado e desligado. Não tem uma sensação de esforço. Então, a capacidade lucida é natural. Então, quando nós removemos os obstáculos, ou seja, nós removamos a ação da mente baseada em referenciais construídos. A lucidez natural se torna a ação natural. Ela surge naturalmente. Então, nesse sentido é que há uma não-ação. Não é necessário qualquer tipo de elaboração ou qualquer tipo de esforço. Isso é uma forma de dizer que não há iluminação, não há sabedoria. Quando a gente olha para a juna parâmita, não há resultado. Não há um caminho para atingir isso. Então, é uma forma de reconhecer essa condição última, com uma condição não elaborada. Se a gente tiver qualquer tipo de esforço para obter alguma coisa, aquilo não é algo natural, aquilo é elaborado. Isso é aquilo é elaborado. Então, mais adiante, vamos nos defrontar com a impermanência inevitável. Aqui o tem um início, tem uma duração, tem um final. Mas o aspecto da lucidez pertence a dar uma caia, dar mata, não há. Nem uma artificialidade, nem o suportanto, não tem nenhum momento em que isso não estivesse presente. Essa realização. Essa realização não surge como um conjunto de palavras ou um estados mentais específicos. Ela surge como a base primordial sem conteúdo, insescentemente presente. Aqui, nós acessamos isso na forma do prática parâmito como uma confiança. E dependendo dos ensinamentos, nós podemos ouvir isso também como o fé. Eu não gosto muito à tradução como o fé, porque parece que é uma ação. Eu tenho que desenvolver uma fé. Mas essa fé é mais parecida com a presença, a natural presença, simplesmente repousar sem esforço na natural presença. Agora, no cotidiano, o que acontece com o nosso? Nós estamos sob o domínio de o padana. Nesse momento, nós estamos inevitabilmente presos ao padana. O padana significa o que? Nós temos que fazer esforço o tempo todo. Nós estamos, o padana é o nono dos doze alos. Nós somos equilibristas. Nós estamos constantemente equilibrando as coisas. Nós estamos constantemente equilibrando. Não só uma coisa, mas nós estamos equilibrando muitas diferentes coisas. Tem uma época que usava esse exemplo dos pratos gerando no círculo. Tenho esses equilibristas. O ciclo terminaram. Não sei, não tem mais como botar. Esse exemplo não é isso. Eles botam mais um prato, mais um prato, mais um prato. No último, o primeiro jota o quadro ainda. Ele vai retornar, vai acelerar. Essa é a nossa estação. Vamos girar um prato aqui. Nós estamos constantemente em ação. Nós estamos num ambiente de não ação. O mundo condicionado, ele exige isso. Isso é dos três animais ao segundo. É uma incessante demanda, uma busca incessante. O mundo se escam, os galos se escam. Os seres buscam sempre alguma coisa incessentemente. Nós não temos como evitar isso. Porque esse é o ponto, é a forma de ela que nós trabalhamos em permanência. Então, nós a gente perguntar mais como é que começou isso? Porque todo mundo tem que se mover o tempo todo. É isso, porque nós somos equilibristas. Nós somos seres desequilibrados. Então, nós estamos constantemente em rei que librando. Não há solução. Então, a não ação é uma descoberta. Uma descoberta dentro do mundo da ação. É alguma coisa extraordinária. Os mestres de diferentes tradições, eles também descobriram isso. Não são os mestres do dia. Eles falam da não ação. Algumas mestres, algumas tradições.
