+++ dharma_transcriptions/downloads/Lama Samten poderia falar sobre os sonhos_/subtitles.srt
@@ -1,0 +1,260 @@
+1
+00:00:00,000 --> 00:00:02,000
+ E o nosso corpo?
+
+2
+00:00:02,000 --> 00:00:04,000
+ E o nosso corpo?
+
+3
+00:00:04,000 --> 00:00:06,000
+ E o nosso corpo?
+
+4
+00:00:06,000 --> 00:00:08,000
+ E o nosso corpo?
+
+5
+00:00:08,000 --> 00:00:10,000
+ E o nosso corpo?
+
+6
+00:00:12,000 --> 00:00:18,000
+ Eu acho muito interessante assim, vocês observarem os sonhos, né?
+
+7
+00:00:18,000 --> 00:00:20,000
+ Eu acho super útil assim.
+
+8
+00:00:20,000 --> 00:00:24,000
+ Então nós temos uma dimensão de sabedoria que não desliga.
+
+9
+00:00:24,000 --> 00:00:26,000
+ A mente ela não apaga,
+
+10
+00:00:26,000 --> 00:00:30,000
+ nós estamos constantemente operando em vários níveis.
+
+11
+00:00:30,000 --> 00:00:34,000
+ Nós temos um nível consciente aqui operando com olhos ouvidos,
+
+12
+00:00:34,000 --> 00:00:38,000
+ nariz, língua, tato e a mente a partir do óleo,
+
+13
+00:00:38,000 --> 00:00:42,000
+ mas nós temos a mente, nós temos outros níveis de mente
+
+14
+00:00:42,000 --> 00:00:44,000
+ insercentemente operando.
+
+15
+00:00:44,000 --> 00:00:48,000
+ É como, por exemplo, também as mães, né?
+
+16
+00:00:48,000 --> 00:00:50,000
+ Que elas estão mexendo uma panela,
+
+17
+00:00:50,000 --> 00:00:54,000
+ imporrando um carrinho e falando no telefone
+
+18
+00:00:54,000 --> 00:00:58,000
+ e ainda dando uma ordem pra alguém em outro lugar assim.
+
+19
+00:00:58,000 --> 00:01:00,000
+ Nós temos várias mentes operando.
+
+20
+00:01:00,000 --> 00:01:02,000
+ Essas, nesse caso, ainda são todas conscientes,
+
+21
+00:01:02,000 --> 00:01:06,000
+ mas nós temos mentes que não são mentes que tejam operando
+
+22
+00:01:06,000 --> 00:01:08,000
+ de modo consciente que a gente esteja vendo.
+
+23
+00:01:08,000 --> 00:01:12,000
+ Então de tanto em tanto elas produzem manifestações
+
+24
+00:01:12,000 --> 00:01:17,000
+ e os sonhos, eles são manifestantes de forma que não são
+
+25
+00:01:17,000 --> 00:01:19,000
+ manifestações dessas mentes,
+
+26
+00:01:19,000 --> 00:01:23,000
+ elas são manifestações pictóricas dessas mentes.
+
+27
+00:01:23,000 --> 00:01:27,000
+ Isso não quer dizer que os sonhos sejam verdadeiros,
+
+28
+00:01:27,000 --> 00:01:30,000
+ mas eles revelam,
+
+29
+00:01:30,000 --> 00:01:35,000
+ elas são mais sutis.
+
+30
+00:01:35,000 --> 00:01:39,000
+ O fato de ela ser mais sutis não significa que elas sejam
+
+31
+00:01:39,000 --> 00:01:41,000
+ valorosas sempre.
+
+32
+00:01:41,000 --> 00:01:44,000
+ Eu, uma vez, eu tive essa experiência de
+
+33
+00:01:44,000 --> 00:01:47,000
+ só voo da ex-sezema, naturalmente nós podemos sonhar
+
+34
+00:01:47,000 --> 00:01:51,000
+ com os infanos e ter boas e desce, ou como matar as pessoas.
+
+35
+00:01:51,000 --> 00:01:55,000
+ Uma vez eu encontrei um a pessoa que tinha um grande
+
+36
+00:01:55,000 --> 00:01:59,000
+ sofrimento de si e ele tinha sido um agente americano
+
+37
+00:01:59,000 --> 00:02:05,000
+ na Coreia e ele dizia que tinha um sofrimento curioso,
+
+38
+00:02:05,000 --> 00:02:08,000
+ porque ele diz que tinha encontrado o budismo na Coreia,
+
+39
+00:02:08,000 --> 00:02:12,000
+ mas ele na Coreia tinha só atrasido tristesa e sofrimento
+
+40
+00:02:12,000 --> 00:02:14,000
+ para aquelas pessoas.
+
+41
+00:02:14,000 --> 00:02:18,000
+ Ele diz que ele tinha uma habilidade de sonhar o que
+
+42
+00:02:18,000 --> 00:02:22,000
+ que aconteceu no dia seguinte, como projetar as ações dele,
+
+43
+00:02:22,000 --> 00:02:24,000
+ e no dia seguinte ele tinha muito vezes, ou seja,
+
+44
+00:02:24,000 --> 00:02:27,000
+ matavam muitas pessoas.
+
+45
+00:02:27,000 --> 00:02:30,000
+ Então, é muito como o ventio,
+
+46
+00:02:30,000 --> 00:02:34,000
+ porque nossa mente pode manifestar isso.
+
+47
+00:02:34,000 --> 00:02:38,000
+ Então, nós podemos ter sonhos lucidos,
+
+48
+00:02:38,000 --> 00:02:43,000
+ muito sentido de lucidez positiva e podemos ter sonhos
+
+49
+00:02:43,000 --> 00:02:46,000
+ que só vão nos afundar em sofrimentos.
+
+50
+00:02:46,000 --> 00:02:49,000
+ Mas nós temos essa mente.
+
+51
+00:02:49,000 --> 00:02:54,000
+ Nós podemos, por exemplo, conduzir essa mente também.
+
+52
+00:02:54,000 --> 00:02:57,000
+ Então, antes de dormir, vocês estudam um pouco,
+
+53
+00:02:57,000 --> 00:03:01,000
+ vocês meditam um pouco, estudam um pouco e dormem com uma pergunta.
+
+54
+00:03:01,000 --> 00:03:04,000
+ E quando acordarem, vocês quando acordarem,
+
+55
+00:03:04,000 --> 00:03:09,000
+ vocês procurem ver se tem uma resposta para aquela pergunta.
+
+56
+00:03:09,000 --> 00:03:12,000
+ Antes de qualquer morgimento que vocês façam,
+
+57
+00:03:12,000 --> 00:03:15,000
+ se não abram ou olhem, vocês perguntam.
+
+58
+00:03:15,000 --> 00:03:21,000
+ E aquilo, como é que é, e deixam a elaboração surgir.
+
+59
+00:03:21,000 --> 00:03:26,000
+ É muito provável que vocês vão ter soluções muito claras sobre isso.
+
+60
+00:03:26,000 --> 00:03:30,000
+ Se vocês tiver uma solução muito clara ou seja, a solução que vier,
+
+61
+00:03:30,000 --> 00:03:35,000
+ vocês se perguntem, qual é a bolha que está a mente que está respondendo?
+
+62
+00:03:35,000 --> 00:03:38,000
+ Vocês procurem ver qual é a bolha, dão.
+
+63
+00:03:38,000 --> 00:03:45,000
+ Porque o alcance do que é visto está na dimensão da bolha onde ela é assamente tá.
+
+64
+00:03:45,000 --> 00:03:47,000
+ Então, no início, a gente não vê bem a bolha,
+
+65
+00:03:47,000 --> 00:03:54,000
+ mas a contemplação da bolha é a chave essencial para entender os sonhos.
+
