+++ dharma_transcriptions/downloads/Lama Samten poderia falar sobre os sonhos_/transcription.txt
@@ -1,0 +1,1 @@
+ E o nosso corpo? E o nosso corpo? E o nosso corpo? E o nosso corpo? E o nosso corpo? Eu acho muito interessante assim, vocês observarem os sonhos, né? Eu acho super útil assim. Então nós temos uma dimensão de sabedoria que não desliga. A mente ela não apaga, nós estamos constantemente operando em vários níveis. Nós temos um nível consciente aqui operando com olhos ouvidos, nariz, língua, tato e a mente a partir do óleo, mas nós temos a mente, nós temos outros níveis de mente insercentemente operando. É como, por exemplo, também as mães, né? Que elas estão mexendo uma panela, imporrando um carrinho e falando no telefone e ainda dando uma ordem pra alguém em outro lugar assim. Nós temos várias mentes operando. Essas, nesse caso, ainda são todas conscientes, mas nós temos mentes que não são mentes que tejam operando de modo consciente que a gente esteja vendo. Então de tanto em tanto elas produzem manifestações e os sonhos, eles são manifestantes de forma que não são manifestações dessas mentes, elas são manifestações pictóricas dessas mentes. Isso não quer dizer que os sonhos sejam verdadeiros, mas eles revelam, elas são mais sutis. O fato de ela ser mais sutis não significa que elas sejam valorosas sempre. Eu, uma vez, eu tive essa experiência de só voo da ex-sezema, naturalmente nós podemos sonhar com os infanos e ter boas e desce, ou como matar as pessoas. Uma vez eu encontrei um a pessoa que tinha um grande sofrimento de si e ele tinha sido um agente americano na Coreia e ele dizia que tinha um sofrimento curioso, porque ele diz que tinha encontrado o budismo na Coreia, mas ele na Coreia tinha só atrasido tristesa e sofrimento para aquelas pessoas. Ele diz que ele tinha uma habilidade de sonhar o que que aconteceu no dia seguinte, como projetar as ações dele, e no dia seguinte ele tinha muito vezes, ou seja, matavam muitas pessoas. Então, é muito como o ventio, porque nossa mente pode manifestar isso. Então, nós podemos ter sonhos lucidos, muito sentido de lucidez positiva e podemos ter sonhos que só vão nos afundar em sofrimentos. Mas nós temos essa mente. Nós podemos, por exemplo, conduzir essa mente também. Então, antes de dormir, vocês estudam um pouco, vocês meditam um pouco, estudam um pouco e dormem com uma pergunta. E quando acordarem, vocês quando acordarem, vocês procurem ver se tem uma resposta para aquela pergunta. Antes de qualquer morgimento que vocês façam, se não abram ou olhem, vocês perguntam. E aquilo, como é que é, e deixam a elaboração surgir. É muito provável que vocês vão ter soluções muito claras sobre isso. Se vocês tiver uma solução muito clara ou seja, a solução que vier, vocês se perguntem, qual é a bolha que está a mente que está respondendo? Vocês procurem ver qual é a bolha, dão. Porque o alcance do que é visto está na dimensão da bolha onde ela é assamente tá. Então, no início, a gente não vê bem a bolha, mas a contemplação da bolha é a chave essencial para entender os sonhos.
